backend/requirements.txt
# Flask core
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.1

# Database
SQLAlchemy==2.0.36

# AI / ML
scikit-learn==1.5.2
numpy==1.26.4

# HTTP client
requests==2.32.3

# Environment variables
python-dotenv==1.0.1

# Validation
marshmallow==3.22.0

# WSGI server (production)
gunicorn==22.0.0
